start implementing guard

+++ src/Dijkstra.NET/Dijkstra.NET/ShortestPath/Dijkstra.cs
@@ -34,7 +34,7 @@
 
                 if (u.Key == to)
                 {
-                    result.Distance = (int) u.Distance;
+                    result.Distance = u.Distance;
                     break;
                 }
